requirements.txt
Cython==0.20.1
PySDL2==0.9.2
numpy==1.8.1
